% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{assign_styles_to_layers}
\alias{assign_styles_to_layers}
\title{Assign a Specified or Default Style to Layers Based on Layer Name}
\usage{
assign_styles_to_layers(
  from,
  to,
  database = NULL,
  schema = "public",
  layers_to_copy = NULL,
  layer_name = NULL
)
}
\arguments{
\item{from}{A data source for the input styles. This can be:
- A string representing the path to a GeoPackage file.
- A `DBI` database connection object to a PostGIS database, created using [RPostgres::dbConnect()].}

\item{to}{A data destination for the output styles. This can be:
- A string representing the path to a GeoPackage file.
- A `DBI` database connection object to a PostGIS database, created using [RPostgres::dbConnect()].}

\item{database}{A string, database name, only in case the destination is in PostGIS.}

\item{schema}{A string, schema name, only in case the destination is in PostGIS.}

\item{layers_to_copy}{An optional character vector specifying the names of layers
in the destination to which the styles should be applied.
If `NULL` (default), applies the style to all layers.}

\item{layer_name}{An optional string representing the name of the layer from the source
`layer_styles` table whose style should be applied. If `NULL` (default), applies the
first style.}
}
\value{
The updated `layer_styles` table, returned invisibly.
}
\description{
This function copies the `layer_styles` from a source to a destination.
By default, it assigns the first style from the source to all layers in the destination,
but a specific style can be selected based on the layer name from the source.
If no style is specified, the first style in the `layer_styles` table is applied.
}
\examples{
# ex1
source_gpkg <- system.file("extdata", "clc.gpkg", package = "clc")
dest_gpkg <- tempfile(fileext = ".gpkg")

layer <- sf::st_read(source_gpkg, layer = "clc")
sf::st_write(layer, dsn = dest_gpkg, layer = "clc2", append = TRUE)

# Assign styles to all layers using the first style
assign_styles_to_layers(from = source_gpkg, to = dest_gpkg)

# Assign styles to specific layers using the first style
assign_styles_to_layers(from = source_gpkg, to = dest_gpkg,
                        layers_to_copy = "clc2")

# Assign a specific style based on layer name to all layers
assign_styles_to_layers(from = source_gpkg, to = dest_gpkg,
                        layer_name = "clc")

\dontrun{
# ex2
conn <- RPostgres::dbConnect(
  RPostgres::Postgres(),
  dbname = 'exampledb',
  host = 'localhost',
  port = '5432',
  user = 'postgres',
  password = 'postgres'
)
dest_gpkg <- "destination.gpkg"

# Assign styles to all layers using the first style
assign_styles_to_layers(from = conn, to = dest_gpkg)

# Assign styles to specific layers using the first style
assign_styles_to_layers(from = conn, to = dest_gpkg,
                        layers_to_copy = c("layer1", "layer2"))

# Assign a specific style based on layer name to all layers
assign_styles_to_layers(from = conn, to = dest_gpkg,
                        layer_name = "layerX")

RPostgres::dbDisconnect(conn)

# ex3
source_gpkg <- "source.gpkg"
conn <- RPostgres::dbConnect(
  RPostgres::Postgres(),
  dbname = 'exampledb',
  host = 'localhost',
  port = '5432',
  user = 'postgres',
  password = 'postgres'
)

# Assign styles to all layers using the first style
assign_styles_to_layers(from = source_gpkg, to = conn, database = 'exampledb')

# Assign styles to specific layers using the first style
assign_styles_to_layers(from = source_gpkg, to = conn, database = 'exampledb',
                        layers_to_copy = c("layer1", "layer2"))

# Assign a specific style based on layer name to all layers
assign_styles_to_layers(from = source_gpkg, to = conn, database = 'exampledb',
                        layer_name = "layerX")

RPostgres::dbDisconnect(conn)


# ex4
source_conn <- RPostgres::dbConnect(
  RPostgres::Postgres(),
  dbname = 'exampledb1',
  host = 'localhost',
  port = '5432',
  user = 'postgres',
  password = 'postgres'
)
conn <- RPostgres::dbConnect(
  RPostgres::Postgres(),
  dbname = 'exampledb2',
  host = 'localhost',
  port = '5432',
  user = 'postgres',
  password = 'postgres'
)

# Assign styles to all layers using the first style
assign_styles_to_layers(from = source_conn, to = conn, database = 'exampledb2')

# Assign styles to specific layers using the first style
assign_styles_to_layers(from = source_conn, to = conn, database = 'exampledb2',
                        layers_to_copy = c("layer1", "layer2"))

# Assign a specific style based on layer name to all layers
assign_styles_to_layers(from = source_conn, to = conn, database = 'exampledb2',
                        layer_name = "layerX")

RPostgres::dbDisconnect(source_conn)
RPostgres::dbDisconnect(conn)
}

}
