% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc_category.R
\name{clc_category}
\alias{clc_category}
\title{`clc_category` S3 class}
\usage{
clc_category(style, values)
}
\arguments{
\item{style}{A data frame containing a QGIS QML style in the column `styleQML`.}

\item{values}{The set of values used to filter the results; filtering is applied only if this set
is not NULL.}
}
\value{
An object of class `clc_category`.
}
\description{
Create an object of class `clc_category`.
}
\details{
This function creates an object of class `clc_category` from a CLC style definition
and a set of values to filter the original style.
}
\examples{
source_gpkg <- system.file("extdata", "clc.gpkg", package = "clc")
obj <- clc(source_gpkg, "clc")

}
