% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clc_code}
\alias{clc_code}
\title{CORINE Land Cover (CLC) Codes}
\format{
A vector.
}
\usage{
clc_code
}
\description{
Each code represents a specific category at the most detailed level (Level 3) of the CLC system.
}
\keyword{datasets}
