% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{clip_vector}
\alias{clip_vector}
\title{Clip a Vector Layer with a Polygon}
\usage{
clip_vector(vector, polygon)
}
\arguments{
\item{vector}{An `sf` object representing the vector layer to be clipped.}

\item{polygon}{An `sf` object representing the polygon layer used for clipping.}
}
\value{
An `sf` object containing the features of the input `vector` that intersect with the `polygon`.
The output will be in the CRS of the `polygon`, and it will retain all attributes of the input `vector`.
}
\description{
This function clips a vector layer (e.g., points, lines, polygons) using a polygon layer.
It handles CRS (Coordinate Reference System) transformations automatically if necessary,
ensuring the output is in the same CRS as the input polygon.
}
\examples{

vector <- sf::st_as_sf(data.frame(
  id = 1:3,
  geometry = sf::st_sfc(
    sf::st_point(c(0.5, 0.5)),
    sf::st_point(c(1.5, 1.5)),
    sf::st_point(c(2.5, 2.5))
  )
), crs = 4326)

polygon <- sf::st_as_sf(data.frame(
  id = 1,
  geometry = sf::st_sfc(sf::st_polygon(list(rbind(
    c(0, 0), c(2, 0), c(2, 2), c(0, 2), c(0, 0)
  ))))
), crs = 4326)

cv <- clip_vector(vector, polygon)

}
