% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigugr_styles.R
\name{copy_styles_layer}
\alias{copy_styles_layer}
\title{Copy styles layer from one source `GeoPackage` to another}
\usage{
copy_styles_layer(from, to)
}
\arguments{
\item{from}{A GeoPackage file name.}

\item{to}{A GeoPackage file name.}
}
\value{
`obj`, invisibly.
}
\description{
Assigns the first style from the source to all layers in the destination.
}
\examples{
#

}
