% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc_category.R
\name{get_colors}
\alias{get_colors}
\alias{get_colors.clc_category}
\title{Retrieve colors from a `clc_category` object}
\usage{
get_colors(cloc)

\method{get_colors}{clc_category}(cloc)
}
\arguments{
\item{cloc}{A `clc_category` object.}
}
\value{
A character vector of colors.
}
\description{
This generic function extracts the color values associated with a
`clc_category` object. It returns a character vector containing the
`color` field from the object.
}
\examples{
#

}
