% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc_category.R
\name{get_levels}
\alias{get_levels}
\alias{get_levels.clc_category}
\title{Retrieve levels from a `clc_category` object}
\usage{
get_levels(cloc)

\method{get_levels}{clc_category}(cloc)
}
\arguments{
\item{cloc}{A `clc_category` object.}
}
\value{
A data frame with columns:
  - `id`: The identifier of the category.
  - `description`: A textual description of the category.
  - `color`: The color associated with the category.
}
\description{
This generic function extracts the levels values associated with a
`clc_category` object. It returns a data frame contains the fields `id`,
`description`, and `color` from the `clc_category` object.
}
\examples{
#

}
