% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc.R
\name{get_spatial_resolution}
\alias{get_spatial_resolution}
\alias{get_spatial_resolution.clc}
\title{Get spatial resolutions}
\usage{
get_spatial_resolution(sr)

\method{get_spatial_resolution}{clc}(sr)
}
\arguments{
\item{sr}{A `clc` object.}
}
\value{
A vector of strings.
}
\description{
Returns the spatial resolutions of the multi-band raster that make up the object.
}
\examples{

esa <- system.file("extdata", "esa", package = "clc")
sr <- clc(dir = esa)

r <- sr |>
     get_spatial_resolution()

}
\seealso{
\code{\link{sat_untarzip}}

Other satellite definition: 
\code{\link{clc}()},
\code{\link{get_band_names}()},
\code{\link{get_spectral_band_names}()}
}
\concept{satellite definition}
