% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_word_counts.R
\name{get_word_counts}
\alias{get_word_counts}
\title{Get Word Frequency Counts}
\usage{
get_word_counts(id = 10, to_file = FALSE, add_index = TRUE)
}
\arguments{
\item{to_file}{Should the output be written to disk as a CSV file? Default is FALSE.}

\item{add_index}{Should there be a column index? Default is TRUE.}

\item{file_id}{Pass in a file_id. Default is 10.}
}
\description{
This function allows you to get word frequency counts from a text file.
}
\examples{
get_word_counts()
}
\keyword{words}
