% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc.R, R/clc_raster.R
\name{plot_clc}
\alias{plot_clc}
\alias{plot_clc.clc}
\alias{plot_clc.clc_raster}
\title{Plot CLC Data}
\usage{
plot_clc(clo, ...)

\method{plot_clc}{clc}(clo, ...)

\method{plot_clc}{clc_raster}(clo, ...)
}
\arguments{
\item{clo}{An object containing CLC data. This must be an instance of a supported class, such as:
- A vectorial data object (e.g., `clc` object).
- A raster data object (e.g., `clc_raster`).}

\item{...}{Additional arguments passed to the specific method for the object's class.}
}
\value{
A plot displaying the CLC data.
}
\description{
Plot CLC data stored in objects of supported classes. The function
adapts the plot based on the class of the input data (vectorial or raster format).
}
\details{
It allows users to plot objects with a simplified interface, while allowing additional arguments
to be passed to `plot` using `...`.
}
\examples{
source_gpkg <- system.file("extdata", "clc.gpkg", package = "clc")
clo <- clc(source = source_gpkg, layer_name = "clc")

temp_file <- tempfile(fileext = ".png")
png(filename = temp_file, width = 800, height = 600)

clo |>
  plot_clc()

dev.off()

}
