% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clip.R
\name{safe_clip_multipolygon}
\alias{safe_clip_multipolygon}
\title{Safely Clip a Multipolygon Vector Layer}
\usage{
safe_clip_multipolygon(vector, polygon)
}
\arguments{
\item{vector}{A `sf` multipolygon vector layer to be clipped.}

\item{polygon}{A `sf` polygon layer used as the clipping geometry.}
}
\value{
A `sf` vector layer with the clipped geometries.
}
\description{
This function clips a `MULTIPOLYGON` vector layer using a polygon layer, handling specific
issues that might arise with geometries encoded incorrectly or containing unknown WKB types.
It serves as a fallback when the `clip_vector()` function fails due to errors like
`ParseException: Unknown WKB type 12`, which is associated with *MULTIPOLYGON* types.
}
\details{
The function ensures that the input layer is correctly encoded as `MULTIPOLYGON` and
uses GDAL utilities for re-encoding if necessary. The output is projected to the CRS
of the clipping polygon.

This solution is inspired by a discussion on handling WKB type errors in R:
<https://gis.stackexchange.com/questions/389814/r-st-centroid-geos-error-unknown-wkb-type-12>.
}
\examples{

vector <- sf::st_as_sf(data.frame(
  id = 1:2,
  geometry = sf::st_sfc(
    sf::st_polygon(list(rbind(c(0, 0), c(2, 0), c(2, 2), c(0, 2), c(0, 0)))),
    sf::st_polygon(list(rbind(c(1, 1), c(3, 1), c(3, 3), c(1, 3), c(1, 1))))
  )
), crs = 4326)

polygon <- sf::st_as_sf(data.frame(
  id = 1,
  geometry = sf::st_sfc(sf::st_polygon(list(rbind(
    c(1, 1), c(2, 1), c(2, 2), c(1, 2), c(1, 1)
  ))))
), crs = 4326)

sc <- safe_clip_multipolygon(vector, polygon)

}
\seealso{
Other transformation functions: 
\code{\link{extract_categories_from_style}()}
}
\concept{transformation functions}
