% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc.R
\name{save_to}
\alias{save_to}
\alias{save_to.clc}
\title{Save a CLC layer and its style to a GeoPackage or PostGIS database}
\usage{
save_to(clo, to, database, schema, layer_name)

\method{save_to}{clc}(clo, to, database = NULL, schema = "public", layer_name = NULL)
}
\arguments{
\item{clo}{A `clc` object.}

\item{to}{A data destination for the output. This can be:
- A string representing the path to a GeoPackage file.
- A `DBI` database connection object to a PostGIS database, created using [RPostgres::dbConnect()].}

\item{database}{A string, database name, only in case the destination is in PostGIS.}

\item{schema}{A string, schema name, only in case the destination is in PostGIS.
Defaults to `'public'`.}

\item{layer_name}{A character string specifying the name of the layer in the output.
If `NULL`, the name of the input `layer` is used.}
}
\value{
clo A `clc` object.
}
\description{
This function saves a CLC layer and its style to a GeoPackage file or a PostGIS database.
The destination is determined by the `to` argument.
}
\details{
The function overwrites the table if it already exists.
}
\examples{

esa <- system.file("extdata", "esa", package = "clc")
clo <- clc(dir = esa)

r <- clo |>
     save_to()

}
