% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{vector_to_raster_layers}
\alias{vector_to_raster_layers}
\title{Convert a Vector Layer to Raster Format}
\usage{
vector_to_raster_layers(
  vector_layer,
  field,
  base_raster = NULL,
  resolution = NULL
)
}
\arguments{
\item{vector_layer}{A vector layer in `sf` format to be rasterized.}

\item{field}{The field in the vector layer used to assign values in the raster.}

\item{base_raster}{(Optional) A raster object to use as the base for rasterization.}

\item{resolution}{(Optional) Numeric resolution to define the raster grid if `base_raster` is not provided.}
}
\value{
A raster object created from the vector layer.
}
\description{
Converts a vector layer (in sf format) into a raster layer using a specified field
and either a base raster or a resolution.
}
\details{
If a base raster is used, the vector layer is cropped to the bounding box of the
base raster, and the resulting raster has the same CRS as the base raster.

The function requires either `base_raster` or `resolution` to be provided.
If both are missing, an error is raised.
}
\examples{
gpkg_path <- system.file("extdata", "clc.gpkg", package = "clc")
vector_layer <- sf::st_read(gpkg_path, layer = 'clc', quiet = TRUE)

raster_path <- system.file("extdata", "mdt.tif", package = "clc")
base_raster <- terra::rast(raster_path)

# Ex1
raster_result <- vector_to_raster_layers(
  vector_layer = vector_layer,
  field = "CODE_18",
  base_raster = base_raster
)

# Ex2
raster_result <- vector_to_raster_layers(
  vector_layer = vector_layer,
  field = "CODE_18",
  resolution = 50
)

}
