% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster.R
\name{vector_to_raster_pg}
\alias{vector_to_raster_pg}
\title{Convert a PostGIS Vector Layer to Raster Format}
\usage{
vector_to_raster_pg(
  conn,
  layer_name,
  schema = "public",
  base_raster = NULL,
  resolution = NULL
)
}
\arguments{
\item{conn}{A database connection object to the source PostGIS database (an active `DBI` connection).}

\item{layer_name}{A string representing the name of the vector layer in the PostGIS database to be converted to raster.}

\item{schema}{An optional string specifying the schema in the PostGIS database where the vector layer is located.
Default is `"public"`.}

\item{base_raster}{An optional `terra` raster object to use as the base for the output raster.
If `NULL`, the function generates a new raster based on the extent of the vector layer and the provided resolution.}

\item{resolution}{An optional numeric value specifying the resolution of the raster (in meters).
This parameter is used to create a new raster if `base_raster` is not provided.}
}
\value{
A `terra` raster object representing the converted vector layer in raster format.
}
\description{
This function converts a vector layer stored in a PostGIS database into a raster format.
It allows the use of an existing raster as a base for the output or, if no base raster is provided,
generates a new raster using the specified resolution and the extent of the vector layer.
}
\details{
The function queries the specified layer from the PostGIS database using the provided `conn`, `layer_name`,
  and `schema` parameters. If a `base_raster` is provided, the output raster will have the same resolution, extent,
  and coordinate reference system as the base raster. If `base_raster` is not provided, the function generates a new
  raster using the specified `resolution` and the extent of the vector layer.
}
\examples{
\dontrun{
# Connect to a PostGIS database
conn <- DBI::dbConnect(
  RPostgres::Postgres(),
  dbname = "mydb",
  user = "myuser",
  password = "mypassword",
  host = "localhost"
)

# Convert a PostGIS vector layer to raster with a specified resolution
raster_result <- vector_to_raster_pg(
  conn = conn,
  layer_name = "land_cover",
  schema = "public",
  resolution = 100
)

# Convert a PostGIS vector layer using an existing raster as base
base_raster <- terra::rast("base_raster.tif")
raster_result_base <- vector_to_raster_pg(
  conn = conn,
  layer_name = "land_cover",
  schema = "public",
  base_raster = base_raster
)

# Disconnect from the database
DBI::dbDisconnect(conn)
}

}
