% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styles.R
\name{assign_styles_to_layers}
\alias{assign_styles_to_layers}
\title{Assign a Specified or Default Style to Layers in a GeoPackage Based on Layer Name}
\usage{
assign_styles_to_layers(from, to, layers_to_copy = NULL, layer_name = NULL)
}
\arguments{
\item{from}{A data source for the input styles. This can be:
- A string representing the path to a GeoPackage file.
- A `DBI` database connection object to a PostGIS database, created using [RPostgres::dbConnect()].}

\item{to}{A string representing the path to the destination GeoPackage file.}

\item{layers_to_copy}{An optional character vector specifying the names of layers
in the destination GeoPackage to which the styles should be applied.
If `NULL` (default), applies the style to all layers.}

\item{layer_name}{An optional string representing the name of the layer from the source
`layer_styles` table whose style should be applied. If `NULL` (default), applies the
first style.}
}
\value{
The updated `layer_styles` table, returned invisibly.
}
\description{
This function copies the `layer_styles` from a source GeoPackage to a destination GeoPackage.
By default, it assigns the first style from the source to all layers in the destination,
but a specific style can be selected based on the layer name from the source GeoPackage.
If no style is specified, the first style in the `layer_styles` table is applied.
}
\examples{
\dontrun{
source_gpkg <- "source.gpkg"
dest_gpkg <- "destination.gpkg"

# Assign styles to all layers using the first style
assign_styles_to_layers(from = source_gpkg, to = dest_gpkg)

# Assign styles to specific layers using the first style
assign_styles_to_layers(from = source_gpkg, to = dest_gpkg,
                        layers_to_copy = c("layer1", "layer2"))

# Assign a specific style based on layer name to all layers
assign_styles_to_layers(from = source_gpkg, to = dest_gpkg,
                        layer_name = "layerX")

conn <- RPostgres::dbConnect(
  RPostgres::Postgres(),
  dbname = 'exampledb',
  host = 'localhost',
  port = '5432',
  user = 'postgres',
  password = 'postgres'
)
dest_gpkg <- "destination.gpkg"

# Assign styles to all layers using the first style
assign_styles_to_layers(from = conn, to = dest_gpkg)

# Assign styles to specific layers using the first style
assign_styles_to_layers(from = conn, to = dest_gpkg,
                        layers_to_copy = c("layer1", "layer2"))

# Assign a specific style based on layer name to all layers
assign_styles_to_layers(from = conn, to = dest_gpkg,
                        layer_name = "layerX")

RPostgres::dbDisconnect(conn)
}

}
