% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc.R
\name{clc}
\alias{clc}
\title{`clc` S3 Class}
\usage{
clc(source, layer_name)
}
\arguments{
\item{source}{The source of the vector layer. Can be a file path to a GeoPackage or a PostGIS connection.}

\item{layer_name}{The name of the layer in the source to be used.}
}
\value{
An object of class `clc`.
}
\description{
Create an object of class `clc`.
}
\details{
This function creates an object of class `clc` from a vector layer in either
a GeoPackage or a PostGIS database.

The layer must have a style defined in the source.
}
\examples{
# ex1
source_gpkg <- system.file("extdata", "clc.gpkg", package = "clc")
clo <- clc(source = source_gpkg, layer_name = "clc")

\dontrun{
# ex2
conn <- RPostgres::dbConnect(
  RPostgres::Postgres(),
  dbname = 'exampledb',
  host = 'localhost',
  port = '5432',
  user = 'postgres',
  password = 'postgres'
)
clo <- clc(source = conn, layer_name = "clc")
}
}
