% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc_raster.R
\name{clc_raster}
\alias{clc_raster}
\title{`clc_raster` S3 class}
\usage{
clc_raster(
  vector_layer,
  field,
  category,
  base_raster = NULL,
  resolution = NULL
)
}
\arguments{
\item{vector_layer}{A vector layer in `sf` format to be rasterized.}

\item{field}{The field in the vector layer used to assign values in the raster.}

\item{category}{A `clc_category` object.}

\item{base_raster}{(Optional) A raster object to use as the base for rasterization.}

\item{resolution}{(Optional) Numeric resolution to define the raster grid if `base_raster` is not provided.}
}
\value{
An object of class `clc_raster`.
}
\description{
Create an object of class `clc_raster`.
}
\details{
This function creates an object of class `clc_raster` from a `terra` raster object
and a `clc_category` object.

The function requires either `base_raster` or `resolution` to be provided.
If both are missing, an error is raised.
}
