% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc_raster.R
\name{get_raster}
\alias{get_raster}
\alias{get_raster.clc_raster}
\title{Retrieve a Raster Representation of CLC}
\usage{
get_raster(clo)

\method{get_raster}{clc_raster}(clo)
}
\arguments{
\item{clo}{A `clc_raster` object.}
}
\value{
A `terra::SpatRaster` object.
}
\description{
Retrieve a raster representation (`terra::SpatRaster`) from a CLC object.
}
\examples{
#

}
