% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc.R, R/clc_raster.R
\name{plot_clc}
\alias{plot_clc}
\alias{plot_clc.clc}
\alias{plot_clc.clc_raster}
\title{Plot CORINE Land Cover Data}
\usage{
plot_clc(clo, ...)

\method{plot_clc}{clc}(clo, ...)

\method{plot_clc}{clc_raster}(clo, ...)
}
\arguments{
\item{clo}{An object containing CLC data. This must be an instance of a supported class, such as:
- A vectorial data object (e.g., `sf` object with polygons).
- A raster data object (e.g., `terra::SpatRaster`).}

\item{...}{Additional arguments passed to the specific method for the object's class.}
}
\value{
A plot displaying the CLC data.
}
\description{
A generic function to plot CORINE Land Cover (CLC) data stored in objects of supported classes.
The function adapts the plot based on the class of the input data (vectorial or raster format).
}
\details{
It allows users to plot objects with a simplified interface, while allowing additional arguments
to be passed to `plot` using `...`.


The function dispatches to specific methods depending on the class of `clo`. It supports:
- Vectorial CLC data, where it plots polygons with their associated styles or categories.
- Raster CLC data, where it visualizes the raster values with a legend based on categories.
}
\examples{
#

}
