% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc.R
\name{prepare_plot}
\alias{prepare_plot}
\alias{prepare_plot.clc}
\title{Prepare a Plot for CLC Data}
\usage{
prepare_plot(clo)

\method{prepare_plot}{clc}(clo)
}
\arguments{
\item{clo}{A `clc` object.}
}
\value{
A `ggplot2` object ready for rendering.
}
\description{
Generates a `ggplot2` object to visualize CLC data. The function
processes the data stored in a `clc` object, ensuring that the specified field is
mapped correctly to the categories and their associated styles.
}
\examples{
#

}
